created DetailedEventCard component to be used in EventDetails & Collab pages
