+++ front-end/src/pages/EventDetails.jsx
@@ -6,29 +6,21 @@ import { getEventDetails, setUserAttending } from "../utilities/EventUtilities";
 import { getUserProfile } from "../utilities/UserProfileUtilities";
 import { getEventIcon } from '../utilities/DefaultIconsUtilities';
 import { getiCalEventDetails } from "../utilities/EventUtilities";
+import gps from "../assets/gps.jpg";
+import DetailedEventCard from "../components/DetailedEventCard";
+
 
 export default function EventDetails() {
   let { eventID } = useParams();
   const [iCalDetails, setiCalDetails] = useState([]);
   const [eventDetails, setEventDetails] = useState([]);
-  const [collaborators, setCollaborators] = useState([]);
   const [usersAttending, setUsersAttending] = useState([]);
   const [eventsAttending, setEventsAttending] = useState([]);
-  const [eventIcon, setEventIcon] = useState("");
   const [userID, setUserID] = useState();
   const myOutletContextObj = useOutletContext();
+
   const { user } = myOutletContextObj;
-  const collaboratorsStr = collaborators.join(", ");
 
-  // Fetches default event icon
-  const fetchEventIcon = async () => {
-    const icon = await getEventIcon()
-    setEventIcon(icon)
-}
-
-  useEffect(() => {
-    fetchEventIcon()
-  }, []);
 
   // Gets list of events user is attending; Used in isUserAttending function
   const handleUserEventsAttending = async () => {
@@ -41,29 +33,10 @@ export default function EventDetails() {
   const getiCalInfo = async () => {
     const response = await getiCalEventDetails(eventID);
     setiCalDetails(response);
-    console.log("EVENT DETAILS page--iCal details:", iCalDetails);
+    // console.log("EVENT DETAILS page--iCal details:", iCalDetails);
 
   };
 
-  // //adding function to format date in more readable way
-  // function switchDateFormat(dateStr) {
-  //   const dateArr = dateStr.split("-");
-  //   let formattedDate = dateArr[1] + "/" + dateArr[2] + "/" + dateArr[0];
-  //   return formattedDate;
-  // }
-  // //adding function to format time in 12-hr format
-  // function switchTimeFormat(timeStr){
-  //   const [hours, minutes, pmam] = timeStr.split(/[:\s]/)
-  //   //convert hours to 12hr if 00,  then returns 12
-  //   const hours12format = (parseInt(hours, 10) % 12 || 12)
-  //   const formattedTime = `${hours12format}:${minutes} ${pmam}`
-  //   return formattedTime
-  // }
-
-
-  useEffect(() => {
-    handleUserEventsAttending();
-  }, []);
 
   // get event details using event utilities and set the event details
   const getEvent = async () => {
@@ -72,15 +45,18 @@ export default function EventDetails() {
     console.log("EVENT DETAILS page--event details:", eventDetails);
 
     setUsersAttending(eventDetails.users_attending);
-    // map through collaborators to get their display names
-    let collabArr = eventDetails.collaborators;
-    let collaborators = collabArr.map((collab) => collab.display_name);
-    setCollaborators(collaborators);
-    console.log(eventDetails.data);
+    //---this is being handled on DetailedEventCard, leaving commented out in case we need it on this page later
+      // map through collaborators to get their display names
+      // let collabArr = eventDetails.collaborators;
+      // let collaborators = collabArr.map((collab) => collab.display_name);
+      // setCollaborators(collaborators);
+      // console.log(eventDetails.data);
   };
 
+  //consolidated useEffects on page
   useEffect(() => {
     getEvent();
+    handleUserEventsAttending();
     getiCalInfo();
   }, []);
 
@@ -104,7 +80,7 @@ export default function EventDetails() {
     } else {
       return false
     }
-  }
+  };
 
   // Renders button conditionally based on if user is RSVPed
   const renderRSVPButton = () => {
@@ -129,139 +105,136 @@ export default function EventDetails() {
                 RSVP
               </Button>
     }
-  }
-
+  };
+  const cardCSS = {width: "35rem"}
   return (
     <Container>
-      <Row className="ustify-content-md-center">
-        <Col className="justify-content-md-center">
+      <Row>
+        <Col>
           <br />
-          <Card style={{ width: "35rem" }} border="light">
-            <Card.Body>
-              <Card.Title
-                as="h1"
-                style={{ fontWeight: "bold", color: "#6840DF" }}
+          {eventDetails && (
+            <DetailedEventCard eventDetails={eventDetails} cardCSS={cardCSS}></DetailedEventCard>
+          )}
+          {renderRSVPButton()}
+        </Col>
+
+        {/*LOCATION IMG &&&& DIRECTIONS BUTTON */}
+        <Col>
+          <Card style={{ width: "20rem" }} sm={4}>
+            <Card.Img src={gps}></Card.Img>
+            <Card.Body className="text-center">
+              <Button
                 className="text-center"
+                variant="info"
+                as={Link}
+                to="/eventdirections"
               >
-                {eventDetails && eventDetails.title}
-              </Card.Title>
-
-              <Card.Subtitle
-                style={{ fontStyle: "italic" }}
-                className="text-center"
-              >
-                Hosted by: {collaboratorsStr}
-              </Card.Subtitle>
-              {/* Conditional rendering of event photo; If event has photo, render that; If no photo, render default event icon */}
-              <Card.Img
-                variant="top"
-                src={eventDetails.event_photo || eventIcon}
-                style={{ height: "500px" }}
-                alt={`${eventDetails.title}'s photo`}
-              />
-
-              <ListGroup variant="flush">
-                {/* !!updated this information to reflect current variable names */}
-                <ListGroup.Item>
-                  <Card.Text
-                    style={{ textDecoration: "underline", fontSize: "larger" }}
-                  >
-                    Event Details:
-                    
-                  </Card.Text>
-                  <ul>
-                    <li>
-                      <strong> Start: </strong>
-                      {eventDetails &&
-                        eventDetails.startDate &&
-                        eventDetails.startTime &&
-                        `${eventDetails.startDate} at ${eventDetails.startTime}`}
-                    </li>
-                    <li>
-                      <strong> End: </strong>
-                      {eventDetails &&
-                        eventDetails.endDate &&
-                        eventDetails.endTime &&
-                        `${eventDetails.endDate} at ${eventDetails.endTime}`}
-                    </li>
-
-                    <li>
-                      {" "}
-                      <strong> Time Zone: </strong>{" "}
-                      {eventDetails && eventDetails.time_zone}
-                    </li>
-                    <li>
-                      <strong> Virtual or In-Person?: </strong>
-                      {eventDetails && eventDetails.event_type}
-                    </li>
-                    {eventDetails.event_type === "Virtual" ? (
-                      <li>
-                        <strong> Event Link: </strong>
-                        {eventDetails && eventDetails.virtual_event_link}
-                      </li>
-                    ) : (
-                      <>
-                        <li>
-                          <strong> Event Venue: </strong>
-                          {eventDetails && eventDetails.event_venue}
-                        </li>
-                        <li>
-                          <strong>Venue Address: </strong>
-                          {eventDetails && eventDetails.event_venue_address}
-                        </li>
-                      </>
-                    )}
-                  </ul>
-                </ListGroup.Item>
-                <ListGroup.Item as="h6" className="text-left">
-                  <h4 style={{ textDecoration: "underline" }}>
-                    About this event:
-                  </h4>
-                  {eventDetails && eventDetails.description}
-                </ListGroup.Item>
-              </ListGroup>
-              <br />
-              <Card.Body className="text-center">
-                <Button
-                  className="text-center"
-                  variant="info"
-                  as={Link}
-                  to="/eventdirections"
-                >
-                  Get Event Directions
-                </Button>
-                <br />
-                <br />
-                {renderRSVPButton()}
-                <div style={{ height: "50px" }}>
-                  <add-to-calendar-button
-                    style={{ height: "50px" }}
-                    size="5"
-                    label="Add to personal calendar"
-                    // buttonsList
-                    // hideTextLabelButton
-                    // buttonStyle="round"
-                    options="'Apple','Google','iCal','Outlook.com','Microsoft 365','Microsoft Teams','Yahoo'"
-                    name={iCalDetails.title}
-                    //   add a conditional rendering for event_type --> do we still need event link for inperson?
-                    location={
-                      eventDetails.event_type === "Virtual"
-                        ? eventDetails.virtual_event_link
-                        : `${eventDetails.event_venue} - ${eventDetails.event_venue_address}`
-                    }
-                    startDate={iCalDetails.startDate}
-                    endDate={iCalDetails.endDate}
-                    startTime={iCalDetails.startTime}
-                    endTime={iCalDetails.endTime}
-                    timeZone={iCalDetails.time_zone}
-                    description={iCalDetails.description}
-                  ></add-to-calendar-button>
-                </div>
-              </Card.Body>
+                Get Event Directions
+              </Button>
             </Card.Body>
           </Card>
+          <add-to-calendar-button
+            style={{ height: "50px" }}
+            size="5"
+            label="Add to personal calendar"
+            options="'Apple','Google','iCal','Outlook.com','Microsoft 365','Microsoft Teams','Yahoo'"
+            name={iCalDetails.title}
+            location={
+              eventDetails.event_type === "Virtual"
+                ? eventDetails.virtual_event_link
+                : `${eventDetails.event_venue} - ${eventDetails.event_venue_address}`
+            }
+            startDate={iCalDetails.startDate}
+            endDate={iCalDetails.endDate}
+            startTime={iCalDetails.startTime}
+            endTime={iCalDetails.endTime}
+            timeZone={iCalDetails.time_zone}
+            description={iCalDetails.description}
+          ></add-to-calendar-button>
         </Col>
       </Row>
     </Container>
   );
 }
+
+//  <Card.Body>
+//    <Card.Title
+//      as="h1"
+//      style={{ fontWeight: "bold", color: "#6840DF" }}
+//      className="text-center"
+//    >
+//      {eventDetails && eventDetails.title}
+//    </Card.Title>
+
+//    <Card.Subtitle style={{ fontStyle: "italic" }} className="text-center">
+//      Hosted by: {collaboratorsStr}
+//    </Card.Subtitle>
+//    {/* Conditional rendering of event photo; If event has photo, render that; If no photo, render default event icon */}
+//    <Card.Img
+//      variant="top"
+//      src={eventDetails.event_photo || eventIcon}
+//      style={{ height: "500px" }}
+//      alt={`${eventDetails.title}'s photo`}
+//    />
+
+//    <ListGroup variant="flush">
+//      {/* !!updated this information to reflect current variable names */}
+//      <ListGroup.Item>
+//        <Card.Text style={{ textDecoration: "underline", fontSize: "larger" }}>
+//          Event Details:
+//        </Card.Text>
+//        <ul>
+//          <li>
+//            <strong> Start: </strong>
+//            {eventDetails &&
+//              eventDetails.startDate &&
+//              eventDetails.startTime &&
+//              `${eventDetails.startDate} at ${eventDetails.startTime}`}
+//          </li>
+//          <li>
+//            <strong> End: </strong>
+//            {eventDetails &&
+//              eventDetails.endDate &&
+//              eventDetails.endTime &&
+//              `${eventDetails.endDate} at ${eventDetails.endTime}`}
+//          </li>
+
+//          <li>
+//            {" "}
+//            <strong> Time Zone: </strong>{" "}
+//            {eventDetails && eventDetails.time_zone}
+//          </li>
+//          <li>
+//            <strong> Virtual or In-Person?: </strong>
+//            {eventDetails && eventDetails.event_type}
+//          </li>
+//          {eventDetails.event_type === "Virtual" ? (
+//            <li>
+//              <strong> Event Link: </strong>
+//              {eventDetails && eventDetails.virtual_event_link}
+//            </li>
+//          ) : (
+//            <>
+//              <li>
+//                <strong> Event Venue: </strong>
+//                {eventDetails && eventDetails.event_venue}
+//              </li>
+//              <li>
+//                <strong>Venue Address: </strong>
+//                {eventDetails && eventDetails.event_venue_address}
+//              </li>
+//            </>
+//          )}
+//        </ul>
+//      </ListGroup.Item>
+//      <ListGroup.Item className="text-left">
+//        <h4 style={{ textDecoration: "underline" }}>About this event:</h4>
+//        {eventDetails && eventDetails.description}
+//      </ListGroup.Item>
+//    </ListGroup>
+//    <br />
+//    <Card.Body className="text-center">
+//      <br />
+//      <br />
+//    </Card.Body>
+//  </Card.Body>;
